requirement.txt
openai-whisper
nltk
fastapi
python-dotenv
requests
torch
torchaudio
torchvision
streamlit           # For UI (planned)
langchain           # For Q&A (upcoming)
faiss-cpu           # For vector storage (upcoming)
